refactor(desktop): 拆出 extensions 设置面板
将 ExtensionConfigurationPanel 与 ExtensionsSettingsPanel 迁入 settings/panels/。
