+++ apps/desktop/src/components/settings-view.tsx
@@ -121,6 +121,8 @@ import { AgentsSettingsPanel } from "@/components/settings/panels/agents-setting
 import { AppearanceSettingsPanel } from "@/components/settings/panels/appearance-settings-panel";
 import { DeveloperSettingsPanel } from "@/components/settings/panels/developer-settings-panel";
 import { DreamSettingsPanel } from "@/components/settings/panels/dream-settings-panel";
+import { ExtensionConfigurationPanel } from "@/components/settings/panels/extension-configuration-panel";
+import { ExtensionsSettingsPanel } from "@/components/settings/panels/extensions-settings-panel";
 import { NetworksSettingsPanel } from "@/components/settings/panels/networks-settings-panel";
 import { RulesSettingsPanel } from "@/components/settings/panels/rules-settings-panel";
 import { SkillsSettingsPanel } from "@/components/settings/panels/skills-settings-panel";
@@ -592,467 +594,6 @@ function McpRuntimeBadge({ state }: { state: McpServerRuntimeBadgeState }) {
 }
 
 
-function ExtensionConfigurationPanel({
-  item,
-  extensionsBusy,
-  onUpdateExtensionSettings,
-  onUpdateExtensionSecret,
-}: {
-  item: DesktopExtensionListItem;
-  extensionsBusy: boolean;
-  onUpdateExtensionSettings: (request: UpdateExtensionSettingsRequest) => Promise<void>;
-  onUpdateExtensionSecret: (request: UpdateExtensionSecretRequest) => Promise<void>;
-}) {
-  const { t } = useTranslation();
-  const [settingDrafts, setSettingDrafts] = useState<Record<string, string>>({});
-  const [secretDrafts, setSecretDrafts] = useState<Record<string, string>>({});
-
-  const settingDraftKey = (extensionId: string, key: string) => `${extensionId}::setting::${key}`;
-  const secretDraftKey = (extensionId: string, key: string) => `${extensionId}::secret::${key}`;
-
-  const updateSettingDraft = (extensionId: string, key: string, value: string) => {
-    setSettingDrafts((current) => ({
-      ...current,
-      [settingDraftKey(extensionId, key)]: value,
-    }));
-  };
-
-  const updateSecretDraft = (extensionId: string, key: string, value: string) => {
-    setSecretDrafts((current) => ({
-      ...current,
-      [secretDraftKey(extensionId, key)]: value,
-    }));
-  };
-
-  const currentSettingText = (
-    item: DesktopExtensionListItem,
-    key: string,
-    fallback?: string | boolean | number,
-  ): string => {
-    const draft = settingDrafts[settingDraftKey(item.id, key)];
-    if (draft !== undefined) {
-      return draft;
-    }
-
-    const value = item.settingsValues?.[key] ?? fallback;
-    return value === undefined || value === null ? "" : String(value);
-  };
-
-  const hasSettings = Boolean(item.settingsSchema?.length);
-  const hasSecrets = Boolean(item.secretSlots?.length);
-
-  return (
-    <div className="space-y-5">
-      <div className="space-y-1">
-        <h1 className="text-xl font-semibold tracking-tight text-foreground">
-          {item.desktopSettingsPage?.title ?? item.displayName}
-        </h1>
-        <p className="text-sm text-muted-foreground">
-          {item.description ?? t('settings.extensionSettingsDescription', { id: item.id })}
-        </p>
-      </div>
-
-      {!hasSettings && !hasSecrets ? (
-        <div className="rounded-lg border border-border/40 bg-background/80 px-4 py-10 text-center text-sm text-muted-foreground">
-          {t('settings.noExtensionSettings')}
-        </div>
-      ) : (
-        <div className="divide-y divide-border/35 rounded-lg border border-border/40 bg-background/80 px-4 sm:px-5">
-          {item.settingsSchema?.map((setting) => {
-            const fieldKey = settingDraftKey(item.id, setting.key);
-            const currentText = currentSettingText(item, setting.key, setting.defaultValue);
-
-            if (setting.type === "boolean") {
-              const checked = Boolean(
-                item.settingsValues?.[setting.key] ?? setting.defaultValue ?? false,
-              );
-              return (
-                <SettingsRow
-                  key={fieldKey}
-                  label={setting.title}
-                  description={setting.description}
-                  htmlFor={fieldKey}
-                >
-                  <div className="flex justify-end">
-                    <Checkbox
-                      id={fieldKey}
-                      checked={checked}
-                      disabled={extensionsBusy}
-                      onCheckedChange={(value) => {
-                        void onUpdateExtensionSettings({
-                          id: item.id,
-                          values: { [setting.key]: value === true },
-                        });
-                      }}
-                      className="size-5"
-                    />
-                  </div>
-                </SettingsRow>
-              );
-            }
-
-            if (setting.type === "select") {
-              const selected = currentText || String(setting.defaultValue ?? "");
-              return (
-                <SettingsRow
-                  key={fieldKey}
-                  label={setting.title}
-                  description={setting.description}
-                  htmlFor={fieldKey}
-                >
-                  <Select
-                    value={selected}
-                    onValueChange={(value) => {
-                      void onUpdateExtensionSettings({
-                        id: item.id,
-                        values: { [setting.key]: value || null },
-                      });
-                    }}
-                    disabled={extensionsBusy}
-                  >
-                    <SelectTrigger id={fieldKey} className="w-full sm:min-w-[14rem]">
-                      <SelectValue placeholder={setting.placeholder ?? setting.title} />
-                    </SelectTrigger>
-                    <SelectContent>
-                      {setting.options?.map((option) => (
-                        <SelectItem key={`${fieldKey}:${option.value}`} value={option.value}>
-                          {option.label}
-                        </SelectItem>
-                      ))}
-                    </SelectContent>
-                  </Select>
-                </SettingsRow>
-              );
-            }
-
-            return (
-              <SettingsRow
-                key={fieldKey}
-                label={setting.title}
-                description={setting.description}
-                htmlFor={fieldKey}
-              >
-                <div className="flex w-full gap-2 sm:max-w-md">
-                  <DesktopFormInput
-                    id={fieldKey}
-                    shellClassName="min-w-0 flex-1"
-                    value={currentText}
-                    disabled={extensionsBusy}
-                    type={setting.type === "number" ? "number" : "text"}
-                    placeholder={setting.placeholder ?? setting.title}
-                    onChange={(event) => updateSettingDraft(item.id, setting.key, event.target.value)}
-                  />
-                  <Button
-                    type="button"
-                    size="sm"
-                    variant="outline"
-                    disabled={extensionsBusy}
-                    onClick={() => {
-                      const raw = settingDrafts[fieldKey] ?? currentText;
-                      const value =
-                        setting.type === "number"
-                          ? raw.trim().length === 0
-                            ? null
-                            : Number(raw)
-                          : raw.trim().length === 0
-                            ? null
-                            : raw;
-                      void onUpdateExtensionSettings({
-                        id: item.id,
-                        values: { [setting.key]: value },
-                      });
-                    }}
-                  >
-                    {t('common.save')}
-                  </Button>
-                </div>
-              </SettingsRow>
-            );
-          })}
-
-          {item.secretSlots?.map((slot) => {
-            const fieldKey = secretDraftKey(item.id, slot.key);
-            const configured = item.secretStatuses?.find((entry) => entry.key === slot.key)?.configured === true;
-            return (
-              <SettingsRow
-                key={fieldKey}
-                label={slot.title}
-                description={slot.description}
-                htmlFor={fieldKey}
-              >
-                <div className="flex w-full flex-wrap justify-end gap-2 sm:max-w-md">
-                  <Badge variant={configured ? "secondary" : "outline"} className="h-9 px-3 text-muted-foreground">
-                    {configured ? t('settings.configured') : t('settings.notConfigured')}
-                  </Badge>
-                  <DesktopFormInput
-                    id={fieldKey}
-                    shellClassName="min-w-0 flex-1"
-                    type="password"
-                    value={secretDrafts[fieldKey] ?? ""}
-                    disabled={extensionsBusy}
-                    placeholder={configured ? t('settings.enterNewValue') : t('settings.enterSecret')}
-                    onChange={(event) => updateSecretDraft(item.id, slot.key, event.target.value)}
-                  />
-                  <Button
-                    type="button"
-                    size="sm"
-                    variant="outline"
-                    disabled={extensionsBusy}
-                    onClick={() => {
-                      void onUpdateExtensionSecret({
-                        id: item.id,
-                        key: slot.key,
-                        value: secretDrafts[fieldKey] ?? "",
-                      });
-                      updateSecretDraft(item.id, slot.key, "");
-                    }}
-                  >
-                    {t('common.save')}
-                  </Button>
-                  <Button
-                    type="button"
-                    size="sm"
-                    variant="ghost"
-                    disabled={extensionsBusy || !configured}
-                    onClick={() => {
-                      void onUpdateExtensionSecret({
-                        id: item.id,
-                        key: slot.key,
-                        value: "",
-                      });
-                      updateSecretDraft(item.id, slot.key, "");
-                    }}
-                  >
-                    {t('common.clear')}
-                  </Button>
-                </div>
-              </SettingsRow>
-            );
-          })}
-        </div>
-      )}
-    </div>
-  );
-}
-
-function ExtensionsSettingsPanel({
-  snapshot,
-  extensionsBusy,
-  onImportExtension,
-  onDeleteExtension,
-}: Pick<
-  SettingsViewProps,
-  | "snapshot"
-  | "extensionsBusy"
-  | "onImportExtension"
-  | "onDeleteExtension"
->) {
-  const { t } = useTranslation();
-  const [deleteTarget, setDeleteTarget] = useState<DesktopExtensionListItem | null>(null);
-  const inputRef = useRef<HTMLInputElement | null>(null);
-  const items = snapshot?.extensionsList ?? [];
-
-  return (
-    <div className="space-y-4">
-      <input
-        ref={inputRef}
-        type="file"
-        accept=".zip,application/zip"
-        className="hidden"
-        onChange={(event) => {
-          const file = event.target.files?.[0];
-          event.target.value = "";
-          if (!file) {
-            return;
-          }
-
-          void (async () => {
-            try {
-              const archiveBase64 = await fileToBase64(file);
-              await onImportExtension({
-                archiveBase64,
-                fileName: file.name,
-              });
-            } catch {
-              /* runtimeError */
-            }
-          })();
-        }}
-      />
-
-      <div className="flex flex-wrap items-center justify-between gap-3">
-        <div className="min-w-0 flex-1 space-y-1">
-          <div className="flex items-center gap-2">
-            <h1 className="text-xl font-semibold tracking-tight text-foreground">{t('settings.extensionsTitle')}</h1>
-            {snapshot?.extensionsLoading ? (
-              <LoaderCircle className="size-4 animate-spin text-muted-foreground" aria-label={t('common.loading')} />
-            ) : null}
-          </div>
-        </div>
-        <Button
-          type="button"
-          size="sm"
-          className="shrink-0"
-          disabled={extensionsBusy}
-          onClick={() => inputRef.current?.click()}
-        >
-          {t('settings.importZip')}
-        </Button>
-      </div>
-
-      <div className="divide-y divide-border/35 rounded-lg border border-border/40 bg-background/80">
-        {items.length === 0 ? (
-          <p className="px-4 py-10 text-center text-sm text-muted-foreground">{t('settings.noExtensionsInstalled')}</p>
-        ) : (
-          items.map((item) => (
-            <div
-              key={item.id}
-              className="flex flex-col gap-3 px-4 py-4 sm:flex-row sm:items-start sm:justify-between sm:gap-4"
-            >
-              <div className="min-w-0 flex-1 space-y-1.5">
-                <div className="flex flex-wrap items-center gap-2">
-                  <span className="text-sm font-medium text-foreground">{item.displayName}</span>
-                  <Badge variant="secondary" className="text-muted-foreground">
-                    {item.version}
-                  </Badge>
-                  {item.author ? (
-                    <Badge variant="secondary" className="text-muted-foreground">
-                      {item.author}
-                    </Badge>
-                  ) : null}
-                </div>
-                {item.description ? (
-                  <p className="text-xs text-muted-foreground">{item.description}</p>
-                ) : null}
-                <p className="truncate font-mono text-[0.65rem] text-muted-foreground/90" title={item.id}>
-                  {item.id}
-                </p>
-                <p className="text-xs text-muted-foreground">
-                  {t('settings.installedAt')}{formatExtensionInstalledAt(item.installedAtUnixMs)}
-                  {item.archiveFileName ? ` · ${t('settings.source')}${item.archiveFileName}` : ""}
-                </p>
-                {item.activationEvents?.length ? (
-                  <p className="text-xs text-muted-foreground">
-                    activationEvents: {item.activationEvents.join(", ")}
-                  </p>
-                ) : null}
-                {item.desktopCss?.length ? (
-                  <div className="space-y-1 pt-1">
-                    <p className="text-xs font-medium text-foreground">{t('settings.desktopCss')}</p>
-                    {item.desktopCss.map((entry) => (
-                      <div
-                        key={`${item.id}:desktop-css:${entry.path}`}
-                        className="rounded-md border border-border/40 bg-muted/20 px-2.5 py-2"
-                      >
-                        <div className="flex flex-wrap items-center gap-2">
-                          <code className="text-[0.7rem] text-foreground">{entry.path}</code>
-                          {entry.media ? (
-                            <Badge variant="outline" className="text-[0.65rem] text-muted-foreground">
-                              media: {entry.media}
-                            </Badge>
-                          ) : null}
-                        </div>
-                        <p className="mt-1 text-xs text-muted-foreground">
-                          {t('settings.desktopCssDescription')}
-                        </p>
-                      </div>
-                    ))}
-                  </div>
-                ) : null}
-                {item.cliHooks?.length ? (
-                  <div className="space-y-1 pt-1">
-                    <p className="text-xs font-medium text-foreground">{t('settings.cliHooks')}</p>
-                    {item.cliHooks.map((hook, index) => (
-                      <div
-                        key={`${item.id}:cli-hook:${hook.slot}:${index}`}
-                        className="rounded-md border border-border/40 bg-muted/20 px-2.5 py-2"
-                      >
-                        <div className="flex flex-wrap items-center gap-2">
-                          <code className="text-[0.7rem] text-foreground">{hook.slot}</code>
-                          {hook.variant ? (
-                            <Badge variant="outline" className="text-[0.65rem] text-muted-foreground">
-                              variant: {hook.variant}
-                            </Badge>
-                          ) : null}
-                        </div>
-                        <p className="mt-1 text-xs text-muted-foreground">
-                          {t('settings.cliHookDescription')}
-                        </p>
-                      </div>
-                    ))}
-                  </div>
-                ) : null}
-              </div>
-              <div className="flex shrink-0 flex-wrap items-center gap-2 self-start sm:self-center">
-                <Button
-                  type="button"
-                  variant="destructive"
-                  size="sm"
-                  disabled={extensionsBusy}
-                  onClick={() => setDeleteTarget(item)}
-                >
-                  {t('common.delete')}
-                </Button>
-              </div>
-            </div>
-          ))
-        )}
-      </div>
-
-      <Dialog
-        open={deleteTarget !== null}
-        onOpenChange={(open) => {
-          if (!open) {
-            setDeleteTarget(null);
-          }
-        }}
-      >
-        <DialogContent className="sm:max-w-md" showCloseButton>
-          <DialogHeader>
-            <DialogTitle>{t('settings.deleteExtension')}</DialogTitle>
-            <DialogDescription>
-              {t('settings.deleteExtensionConfirm', { name: deleteTarget?.displayName ?? '' })}
-            </DialogDescription>
-          </DialogHeader>
-          <div className="flex flex-col-reverse justify-end gap-2 pt-2 sm:flex-row">
-            <Button
-              type="button"
-              variant="outline"
-              size="sm"
-              onClick={() => setDeleteTarget(null)}
-              disabled={extensionsBusy}
-            >
-              {t('common.cancel')}
-            </Button>
-            <Button
-              type="button"
-              variant="destructive"
-              size="sm"
-              disabled={extensionsBusy || !deleteTarget}
-              onClick={() => {
-                const target = deleteTarget;
-                if (!target) {
-                  return;
-                }
-                void (async () => {
-                  try {
-                    await onDeleteExtension({ id: target.id });
-                    setDeleteTarget(null);
-                  } catch {
-                    /* runtimeError */
-                  }
-                })();
-              }}
-            >
-              {extensionsBusy ? <LoaderCircle className="size-4 animate-spin" /> : null}
-              {t('common.delete')}
-            </Button>
-          </div>
-        </DialogContent>
-      </Dialog>
-    </div>
-  );
-}
 
 const mcpCreateScopeOptions: Array<{
   scope: DesktopMcpScope;
